Disk values can be floats

+++ src/js/stores/AppFormTransforms.js
@@ -1,6 +1,6 @@
 const AppFormTransforms = {
   appId: (id) => id + "/transformed",
-  disk: (value) => parseInt(value),
+  disk: (value) => parseFloat(value),
   env: (rows) => {
     return rows.reduce((memo, row) => {
       memo[row.key] = row.value;
